requirements.txt
certifi            
charset-normalizer 
colorama           
contourpy          
cycler             
filelock          
fonttools          
fsspec             
idna              
Jinja2             
kiwisolver         
MarkupSafe        
matplotlib         
mpmath             
networkx           
numpy              
opencv-python      
packaging         
pandas             
pillow             
pip                
psutil             
py-cpuinfo         
pyparsing          
python-dateutil    
pytz               
PyYAML             
requests           
scipy              
seaborn            
setuptools        
six                
sympy              
torch              
torchvision        
tqdm               
typing_extensions  
tzdata             
ultralytics        
ultralytics-thop   
urllib3          